move toggleToDo, simplify logic (remove cursor variable)
